Enlever les commentaires et remplacement des int du temps pour des doubles

+++ NexusSimulatHeure/src/Metier/Source/SourceFinie.java
@@ -6,7 +6,7 @@ import Metier.DistributionTriangulaire;
 public class SourceFinie extends Source {
     private int nombreMax;
     
-    public SourceFinie(int nombreMax, Point point, int heureDepart, int frequence, DistributionTriangulaire distribution){
+    public SourceFinie(int nombreMax, Point point, double heureDepart, double frequence, DistributionTriangulaire distribution){
         super(point, heureDepart, frequence, distribution);
         this.nombreMax = nombreMax;
     }
